honksquad: fix: move cybernetic ears handler to shared and add deletion guard
Move CanHearAttemptEvent handler from server-only CyberneticOrganEffectsSystem
to shared DeafableSystem to fix client prediction desync (#335).
Add TerminatingOrDeleted guard in UpdateIsDeaf to prevent events during
entity deletion (#337).
